Refactor non template functions to cpp files
